git-svn-id: svn+ssh://marioosh.net/var/svn/python@43 15dd19fa-65d8-43a4-bfc9-50ebb766958d

+++ python_learn/src/net/marioosh/python/learn/manager.py
@@ -52,9 +52,11 @@ try:
                 if reponame != None:
                     repopath = os.path.join(SVNDIR,reponame)
                     if os.path.exists(repopath):
-                        tn = raw_input('deleting repo "'+reponame+'" - are you sure ? (y/n) ? ')
-                        if tn == 'y' or tn == 'Y':
-                            shutil.rmtree(repopath)
+                        stat = os.stat(repopath)
+                        if stat.st_uid == os.geteuid():
+                            tn = raw_input('deleting repo "'+reponame+'" - are you sure ? (y/n) ? ')
+                            if tn == 'y' or tn == 'Y':
+                                shutil.rmtree(repopath)
                     else:
                         print ('error: repo "'+reponame+ '" doesn\'t exist')
             else:
